requirements.txt
google-generativeai>=0.7.2
google-cloud-aiplatform>=1.66.0
langchain>=0.2.11
langgraph>=0.0.63
langchain-google-genai>=1.0.3
python-dotenv>=1.0.1
pydantic>=2.8.2
rich>=13.8.0
networkx>=3.3
pandas>=2.2.2
requests>=2.32.3
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
prometheus-client>=0.20.0
matplotlib>=3.9.2
plotly>=5.23.0
pytest>=8.2.2
pytest-asyncio>=0.23.8

